requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
django-ratelimit>=4.1,<5.0
django-redis>=5.4,<6.0
pandas>=2.0,<3.0
openpyxl>=3.1,<4.0
twilio>=8.0,<9.0
celery>=5.3,<6.0
redis>=5.0,<6.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
